feat: 캘린더 컴포넌트 데이터 미리 불러오기 기능 추가 및 차트 데이터 처리 함수 리팩토링
